+++ .gitignore
@@ -137,3 +137,6 @@ dmypy.json
 
 #Example generated files
 examples/evaluations/text_generation/final_metrics.txt
+
+#Data generated by langfair data_loader module
+langfair/data/*
